GUI to segment elastic beads and calculate tension dipols
This GUI allows to manually locate elastic beads in a stack of tiff-images. After detection of the beads they are automatically segmented and spherical harmonics are fitted to the segmentation. When a connection of matlab to comsol server is established, the tension dipol leading to the segmented deformation of the bead can be calculated.