New version of GUI. Improved speed and data usage.
Major changes in handling of saving and loading tiff-stacks. Both should be much faster now. The BigTiff standard is supported now. Furthermore, all images are saved as 8-bit, which reduces data usage. Also a bug in saving the harmonic displacement field is fixed, which increases speed and reduces data usage dramatically.
Last but not least, two features where added: the interpolation of the input images can now be adjusted. Standard is now no interpolation which makes sense for beads existing of many (>50x50x50) voxels. For smaller ones the interpolation can be increased. Also there is a button for export of the figures now.